backend/src/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
python-dotenv==0.21.1